Split grid and ov manager into class files, fixed tox test configuration
