+++ tests/resources/array.ini
@@ -131,7 +131,7 @@ wd_gradient_params = [[0, 0, 0], [0, 0, 0]]
 wd_stig_params = {}
 array_index = [null, null]
 roi_index = [null, 0]
-array_file = ../array/example/wafer_example1.mass.json
+array_file = array/example/wafer_example1.mass.json
 sw_sh = [[0, 0], [0, 0]]
 
 [overviews]
