+++ tox.ini
@@ -7,7 +7,6 @@ skipsdist = True
 
 [testenv]
 description = run tests
-changedir = src
 setenv =
     PYTHONPATH = {env:PYTHONPATH}{:}src
 passenv =
@@ -24,4 +23,5 @@ deps =
     pytest-sugar
     -r requirements.txt
 commands =
-    pytest {posargs:../tests} --color=yes --cov=. --cov-config=../.coveragerc
+    pytest tests
+    #pytest {posargs:tests} --color=yes --cov=. --cov-config=.coveragerc
